// api/src/extensions/channels/offline/types.ts
export namespace Offline {
// ...
  export enum IncomingMessageType {
    text = 'text',
    quick_reply = 'quick_reply',
    postback = 'postback',
    location = 'location',
    file = 'file',
  }

  export type EventType = Offline.StatusEventType | Offline.IncomingMessageType;

  export enum OutgoingMessageType {
    text = 'text',
    buttons = 'buttons',
    quick_replies = 'quick_replies',
    file = 'file',
    list = 'list',
    carousel = 'carousel',
  }

  export type IncomingTextMessageData = { text: string };
// ...
  export type StatusEvent =
    | StatusDeliveryEvent
    | StatusReadEvent
    | StatusTypingEvent;

  export type IncomingTextMessage = {
    type: IncomingMessageType.text;
    data: IncomingTextMessageData;
  };

  export type IncomingPayloadMessage = {
    type: IncomingMessageType.postback | IncomingMessageType.quick_reply;
    data: IncomingPayloadMessageData;
  };

  export type IncomingLocationMessage = {
    type: IncomingMessageType.location;
    data: IncomingLocationMessageData;
  };

  export type IncomingAttachmentMessage = {
    type: IncomingMessageType.file;
    data: IncomingAttachmentMessageData;
  };

  export type IncomingMessageBase =
    | IncomingTextMessage
    | IncomingPayloadMessage
    | IncomingLocationMessage
    | IncomingAttachmentMessage;

  export type IncomingMessage<
    T =
      | IncomingTextMessage
      | IncomingPayloadMessage
      | IncomingLocationMessage
      | IncomingAttachmentMessage,
  > = T & {
    mid?: string;
    author?: string;
    read?: boolean;
    delivery?: boolean;
    // Whether it's a synchronization
    // This is used when message sent by the chatbot from the client side
    sync?: boolean;
    createdAt?: Date;
  };

  export type Event = IncomingMessage | StatusEvent;
// ...
}